Se modulariza el componente dashboard pasando a componentes los quickactions y las dailyactivity
